Split DiceBox asset cleanup script
