+++ scripts/cleanup-dicebox-assets.cjs
@@ -1,0 +1,20 @@
+#!/usr/bin/env node
+// Removes unused asset copies generated by @3d-dice/dice-box's install script.
+
+const { existsSync, rmSync } = require('fs');
+const { join } = require('path');
+
+const generatedAssetDirs = [
+  'public/assets/ammo',
+  'public/assets/themes',
+];
+
+for (const dir of generatedAssetDirs) {
+  const generatedAssetDir = join(process.cwd(), dir);
+  if (!existsSync(generatedAssetDir)) {
+    continue;
+  }
+
+  rmSync(generatedAssetDir, { recursive: true, force: true });
+  console.log(`[dice-box-assets] removed unused generated ${dir}`);
+}
